some simplifications and inlining of StagedCommitState constructor
